# openapi.yaml
openapi: 3.1.0
info:
  title: Fat Loss Coach Inventory API
  version: 1.0.0
  description: API used by a Custom GPT fat-loss coach to manage Google Sheets inventory, body-weight logs, meal logs, and usage logs.
servers:
  - url: https://YOUR-DEPLOYED-API-DOMAIN
    description: Replace with your deployed FastAPI base URL
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    InventoryItemIn:
      type: object
      required: [item_en, quantity, unit]
      properties:
        item_en: { type: string, description: English item name, example: chicken breast }
        item_zh: { type: string, description: Chinese item name, example: 鸡胸肉 }
        category: { type: string, enum: [protein, carb, veg, fruit, condiment, staple, snack, drink, other], default: other }
        quantity: { type: number, example: 900 }
        unit: { type: string, example: g }
        storage: { type: string, enum: [fridge, freezer, pantry, counter], default: fridge }
        purchase_date: { type: string, description: YYYY-MM-DD, example: '2026-06-11' }
        opened_date: { type: string, description: YYYY-MM-DD or blank }
        use_by: { type: string, description: YYYY-MM-DD or blank }
        calories_per_100g: { type: number }
        protein_per_100g: { type: number }
        carbs_per_100g: { type: number }
        fat_per_100g: { type: number }
        status: { type: string, enum: [active, finished, archived, discarded], default: active }
        priority: { type: string, enum: [high, medium, low], default: medium }
        min_quantity: { type: number }
        target_quantity: { type: number }
        notes: { type: string }
    AddInventoryPayload:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/InventoryItemIn' }
    InventoryUpdate:
      type: object
      required: [id, fields]
      properties:
        id: { type: string, example: inv-abc12345 }
        fields:
          type: object
          additionalProperties: true
          description: Fields to update, e.g. quantity, status, use_by, notes.
    UpdateInventoryPayload:
      type: object
      required: [updates]
      properties:
        updates:
          type: array
          items: { $ref: '#/components/schemas/InventoryUpdate' }
    ConsumeItem:
      type: object
      required: [quantity_used, unit]
      properties:
        item_id: { type: string, description: Preferred identifier for exact match }
        item_en: { type: string, description: Fallback match by English name }
        item_zh: { type: string, description: Fallback match by Chinese name }
        quantity_used: { type: number, example: 300 }
        unit: { type: string, example: g }
        meal: { type: string, example: lunch }
        reason: { type: string, example: actual_consumed }
        notes: { type: string }
    ConsumePayload:
      type: object
      required: [items]
      properties:
        date: { type: string, description: YYYY-MM-DD }
        items:
          type: array
          items: { $ref: '#/components/schemas/ConsumeItem' }
        dry_run: { type: boolean, default: false, description: Validate and project without writing }
    WeightEntry:
      type: object
      required: [weight_lb]
      properties:
        date: { type: string, description: YYYY-MM-DD }
        weight_lb: { type: number, example: 189.2 }
        sleep_hours: { type: number }
        hunger_1_10: { type: number }
        training_perf_1_10: { type: number }
        training: { type: string }
        calories: { type: number }
        protein_g: { type: number }
        fast_food: { type: string, enum: [yes, no] }
        alcohol: { type: string, enum: [yes, no] }
        night_snack: { type: string, enum: [yes, no] }
        notes: { type: string }
    MealEntry:
      type: object
      required: [meal_type]
      properties:
        date: { type: string, description: YYYY-MM-DD }
        meal_type: { type: string, example: lunch }
        planned_or_actual: { type: string, enum: [planned, actual], default: planned }
        calories: { type: number }
        protein_g: { type: number }
        carbs_g: { type: number }
        fat_g: { type: number }
        food_summary: { type: string }
        inventory_used: { type: string }
        notes: { type: string }
    MealLogPayload:
      type: object
      required: [meals]
      properties:
        meals:
          type: array
          items: { $ref: '#/components/schemas/MealEntry' }
paths:
  /health:
    get:
      operationId: health
      summary: Check whether the API is reachable.
      responses:
        '200': { description: OK }
  /setup/headers:
    post:
      operationId: setupHeaders
      summary: Write standard headers to all expected Google Sheet tabs.
      description: Use once after creating/importing the spreadsheet template. Does not delete data rows.
      responses:
        '200': { description: Headers written }
  /coach-context:
    get:
      operationId: getCoachContext
      summary: Get all context needed for a daily meal plan.
      description: Returns settings, active inventory, expiring inventory, weight trend, and suggested shopping list.
      parameters:
        - name: expiring_days
          in: query
          schema: { type: integer, default: 3, minimum: 0, maximum: 14 }
      responses:
        '200': { description: Daily coach context }
  /inventory:
    get:
      operationId: getInventory
      summary: Get current inventory.
      parameters:
        - name: include_archived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200': { description: Inventory items }
  /inventory/add:
    post:
      operationId: addInventory
      summary: Add newly purchased ingredients to inventory.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddInventoryPayload' }
      responses:
        '200': { description: Created inventory rows }
  /inventory/update:
    post:
      operationId: updateInventory
      summary: Update existing inventory rows by ID.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateInventoryPayload' }
      responses:
        '200': { description: Updated inventory rows }
  /inventory/consume:
    post:
      operationId: consumeInventory
      summary: Deduct actual consumed ingredients from inventory and write usage log.
      description: Use only after the user confirms they actually consumed the listed ingredients, or use dry_run=true to preview deductions.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsumePayload' }
      responses:
        '200': { description: Inventory consumption result }
  /inventory/expiring:
    get:
      operationId: getExpiringInventory
      summary: Get ingredients expiring soon.
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 3, minimum: 0, maximum: 30 }
      responses:
        '200': { description: Expiring inventory items }
  /weight-log:
    post:
      operationId: logWeight
      summary: Log morning body weight and related daily signals.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WeightEntry' }
      responses:
        '200': { description: Weight logged }
    get:
      operationId: getWeightLog
      summary: Get recent body-weight log.
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30, minimum: 1, maximum: 365 }
      responses:
        '200': { description: Weight log rows }
  /trend-summary:
    get:
      operationId: getTrendSummary
      summary: Compute 7-day vs previous 7-day body-weight trend.
      responses:
        '200': { description: Trend summary }
  /meal-log:
    post:
      operationId: logMeals
      summary: Log planned or actual meals and macro estimates.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MealLogPayload' }
      responses:
        '200': { description: Meal entries logged }
  /shopping-list/suggested:
    get:
      operationId: getSuggestedShoppingList
      summary: Suggest replenishment based on inventory min/target quantities.
      responses:
        '200': { description: Suggested shopping list }